backend/requirements.txt
boto3>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.0.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
